judy-customer-support-backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
faker
